% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ARIES_csv.R
\name{read_ARIES_csv}
\alias{read_ARIES_csv}
\title{Ingest a csv file from the \href{https://hub.mph.in.gov/dataset/aries-crash-data-2007-2017}{ARIES crash repository}}
\usage{
read_ARIES_csv(ARIES_csv)
}
\arguments{
\item{ARIES_csv}{A csv file originating from the ARIES crash repository}
}
\value{
A tibble
}
\description{
Ingest a csv file from the \href{https://hub.mph.in.gov/dataset/aries-crash-data-2007-2017}{ARIES crash repository}
}
\examples{

read_ARIES_csv("ARIES_2017.csv")
}
