% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_ARIES_dates.R
\name{standardize_ARIES_dates}
\alias{standardize_ARIES_dates}
\title{Ensure all dates in an ARIES dataframe follow the same convention of year-month-day}
\usage{
standardize_ARIES_dates(df)
}
\arguments{
\item{df}{A tibble of data imported using read_ARIES_csv()}
}
\value{
A tibble with standardized dates
}
\description{
Ensure all dates in an ARIES dataframe follow the same convention of year-month-day
}
